Jumped right to the bithacks O(1) solution. http://graphics.stanford.edu/~seander/bithacks.html
I have used this in to count the number of enum flags set in the past,
So I was fimilar with it.

The simplest solution is to loop over the bits and check if each one is
set. This is not too bad most of the time as the search space is normally
only 32 bits.